apps/backend/requirements.txt
# =============================================================================
# TRAIDER V1 - Backend Dependencies
# Institutional-grade Python packages for autonomous trading platform
# =============================================================================

# CORE FRAMEWORK
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.9.2
pydantic-settings==2.7.1
python-multipart==0.0.20

# AUTHENTICATION & SECURITY
PyJWT==2.10.1
python-jose[cryptography]==3.5.0  # RESTORED: Required for JWT handling in auth.py
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==44.0.0

# DATABASE & ORM
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
databases[postgresql]==0.9.0

# LOGGING & MONITORING
structlog==24.2.0  # RESTORED: Required for structured logging in utils/logging.py
loguru==0.7.3
prometheus-client==0.21.1

# TESTING & QUALITY ASSURANCE
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-ruff==0.5.0
hypothesis==6.135.14  # RESTORED: Required for property-based testing in test files
factory-boy==3.3.1

# DEVELOPMENT TOOLS
ruff==0.8.4
mypy==1.10.0
black==24.10.0
isort==5.13.2

# HTTP CLIENT
httpx==0.28.1
requests==2.32.3

# UTILITIES
python-dotenv==1.0.1
click==8.1.8
rich==13.9.4

# ASYNC SUPPORT
anyio==4.7.0

# MARKET DATA & TRADING
ccxt==4.4.37
pandas==2.2.3
numpy==2.2.1

# MACHINE LEARNING
scikit-learn==1.6.0
joblib==1.4.2

# FINANCIAL DATA
yfinance==0.2.48

# CONFIGURATION
pyyaml==6.0.2
toml==0.10.2

# SECURITY SCANNING
bandit==1.8.0
safety==3.5.2
sarif-om==1.0.4
jschema-to-python==1.2.3

# PERFORMANCE
orjson==3.10.12

# TIMEZONE HANDLING
pytz==2024.2

# VALIDATION
email-validator==2.2.0

# CACHING
redis==5.2.1

# ENVIRONMENT MANAGEMENT
python-decouple==3.8

# MONITORING & ALERTING
sentry-sdk==2.20.0

# WEBSOCKETS
websockets==14.1

# SERIALIZATION
marshmallow==3.23.2

# DATETIME UTILITIES
python-dateutil==2.9.0

# STATISTICAL ANALYSIS
scipy==1.15.0

# DATA VALIDATION
cerberus==1.3.5

# ASYNC UTILITIES
aiofiles==24.1.0

# HTTP UTILITIES
urllib3==2.3.0

# PROCESS MONITORING
psutil==6.1.1

# TESTING UTILITIES
responses==0.25.5

# PARALLEL TESTING
pytest-xdist==3.6.0

# COVERAGE REPORTING
coverage-badge==1.1.2

# DOCUMENTATION
sphinx==8.1.3

# JSON UTILITIES
jsonschema==4.23.0

# PROGRESS BARS
tqdm==4.67.1

# COLORED OUTPUT
colorama==0.4.6

# COLLECTIONS
more-itertools==10.5.0

# AIOHTTP
aiohttp==3.10.11 